+++ packages/ui/src/components/base-handle.tsx
@@ -1,7 +1,7 @@
-import type { ComponentProps } from "react";
 import { Handle, type HandleProps } from "@xyflow/react";
+import type { ComponentProps } from "react";
 
-import { cn } from "@workspace/ui/lib/utils";
+import { cn } from "@react-flowkit/ui/lib/utils";
 
 export type BaseHandleProps = HandleProps;