+++ packages/ui/src/components/collapsible.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import * as React from "react";
 import * as CollapsiblePrimitive from "@radix-ui/react-collapsible";
+import * as React from "react";
 
-import { cn } from "@workspace/ui/lib/utils";
+import { cn } from "@react-flowkit/ui/lib/utils";
 
 function Collapsible({
   ...props